fix/wpt-rfid: add keep listeners

+++ src/main/helpers/launch_wpt.js
@@ -10,10 +10,14 @@ module.exports = function launchWpt(wpt, callback) {
 
 	return new Promise((resolve, reject) => {
 		let timeout = setTimeout(() => {
-			child.stdout.removeAllListeners()
-			child.stderr.removeAllListeners()
+			if (child.stdout) {
+				child.stdout.removeAllListeners()
+			}
+			if (child.stderr) {
+				child.stderr.removeAllListeners()
+			}
 			child.removeAllListeners()
-			if (!child.killed) {
+			if (child && !child.killed) {
 				child.kill('SIGKILL')
 			}
 			if (wptPid) {
@@ -63,18 +67,19 @@ module.exports = function launchWpt(wpt, callback) {
 		}
 
 		const options = {
-			stdio: wpt.wait_on_ipc ? ['pipe', 'pipe', 'pipe']:  undefined,
-			// windowsHide: true,
+			stdio: wpt.wait_on_ipc ? ['pipe', 'pipe', 'pipe']: undefined,
+			windowsHide: true,
 			shell:wpt.shell,
 			detached: wpt.detached,
 		}
 
-		if (wpt.wait_on_ipc && isScript && path.extname(exePath) === '.sh' || isJs) {
+		if (wpt.wait_on_ipc && options.stdio && isScript && (path.extname(exePath) === '.sh' || isJs)) {
 			// not working on Windows with .bat ...
 			options.stdio.push('ipc')
 		}
 
 		const child = spawn(exe, args, options)
+
 		if (wpt.wait_on_ipc) {
 			child.on('message', message => {
 				log.info('wpt.send', message)
@@ -91,8 +96,12 @@ module.exports = function launchWpt(wpt, callback) {
 						clearTimeout(timeout)
 						timeout = null
 					}
-					child.stdout.removeAllListeners()
-					child.stderr.removeAllListeners()
+					if (child.stdout) {
+						child.stdout.removeAllListeners()
+					}
+					if (child.stderr) {
+						child.stderr.removeAllListeners()
+					}
 					child.removeAllListeners()
 					resolve(child)
 				}
@@ -100,8 +109,8 @@ module.exports = function launchWpt(wpt, callback) {
 		}
 
 		if (
-			!wpt.wait_on_ipc ||
-			(process.env.DEBUG && process.env.DEBUG === 'wpt')
+			child.stdout && (!wpt.wait_on_ipc ||
+			(process.env.DEBUG && process.env.DEBUG === 'wpt'))
 		) {
 			child.stdout.on('data', function (data) {
 				if (process.env.DEBUG && process.env.DEBUG === 'wpt') {
@@ -121,16 +130,20 @@ module.exports = function launchWpt(wpt, callback) {
 							break
 						}
 					}
-					pid = pid.split(" ")
+					const pids = pid.split(" ")
 					if (pid.length > 0) {
-						pid = pid.pop()
+						pid = pids.pop()
 						pid = Number.parseInt(pid, 10)
 						if (Number.isNaN(pid)) {
-							pid = pid.pop()
-							pid = Number.parseInt(pid, 10)
+							if (pids.length > 0) {
+								pid = pids.pop()
+								pid = Number.parseInt(pid, 10)
+							} else {
+								pid = null
+							}
 						}
 
-						if (!Number.isNaN(pid) && callback) {
+						if (pid && !Number.isNaN(pid) && callback) {
 							callback('get_wpt_pid_done', pid)
 						}
 					}
@@ -153,38 +166,20 @@ module.exports = function launchWpt(wpt, callback) {
 				}
 			})
 		}
-
-		child.stderr.on('data', function (data) {
-			//   if (messages.length === 0) {
-			//     setTimeout(() => {
-			//       child.kill('SIGKILL')
-			//       if (wptPid) {
-			//         try {
-			//           process.kill(wptPid)
-			//         } catch (err2) {
-			//           // console.log(err2)
-			//         }
-			//       }
-			//       child.stdout.removeAllListeners()
-			//       child.stderr.removeAllListeners()
-			//       child.removeAllListeners()
-			//       const err = new CustomError(
-			//         400,
-			//         CustomError.CODE.WPT_CREATION_FAILED,
-			//         wpt.path,
-			//         []
-			//       )
-			//       err.messages = messages
-			//       reject(err)
-			//     }, 1000)
-			//   }
-			messages += data.toString()
-		})
-
+		if (child.stderr) {
+			child.stderr.on('data', function (data) {
+				messages += data.toString()
+			})
+		}
+		
 		child.once('exit', reason => {
 			setTimeout(() => {
-				child.stdout.removeAllListeners()
-				child.stderr.removeAllListeners()
+				if (child.stdout) {
+					child.stdout.removeAllListeners()
+				}
+				if (child.stderr) {
+					child.stderr.removeAllListeners()
+				}
 				child.removeAllListeners()
 				reject(
 					new CustomError(
@@ -205,16 +200,33 @@ module.exports = function launchWpt(wpt, callback) {
 			}
 			if (!child.killed) {
 				child.kill('SIGKILL')
-				child.stdout.removeAllListeners()
-				child.stderr.removeAllListeners()
+				if (child.stdout) {
+					child.stdout.removeAllListeners()
+				}
+				if (child.stderr) {
+					child.stderr.removeAllListeners()
+				}
 				child.removeAllListeners()
 			}
 			err.messages = messages
 			reject(err)
 		})
 
-		// child.stdout.removeAllListeners()
-		// child.stderr.removeAllListeners()
-		// resolve(child)
+		if (child && wpt.shell) {
+			if (timeout) {
+				clearTimeout(timeout)
+				timeout = null
+			}
+			setTimeout(() => {
+				if (child.stdout) {
+					child.stdout.removeAllListeners()
+				}
+				if (child.stderr) {
+					child.stderr.removeAllListeners()
+				}
+				child.removeAllListeners()
+				resolve(child)
+			}, 3000)
+		}
 	})
 }
